Palette view for pixel format,
in spirit of SDL wiki, that discourages manual creation of palettes.
Good enough for simple use cases, but in general would be better to
implement a smart pointer wrapper over the SDL_Palette object, similar
to surface and pixel format, and make sure the palette is properly
set/copied when creating new surfaces with the format that holds it.